refactor: Position cell 패키지로 이동
