+++ src/main/kotlin/minesweeper/domain/cell/Empty.kt
@@ -1,5 +1,3 @@
 package minesweeper.domain.cell
 
-import minesweeper.domain.common.Position
-
 class Empty(position: Position, var numberOfNearbyMines: Int = 0) : Cell(position)
