+++ src/main/kotlin/minesweeper/domain/cell/Mine.kt
@@ -1,5 +1,3 @@
 package minesweeper.domain.cell
 
-import minesweeper.domain.common.Position
-
 class Mine(position: Position) : Cell(position)
